backend/requirements.txt
# Django and extensions
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0

# Production server
gunicorn>=21.2
whitenoise>=6.6

# Database
psycopg2-binary>=2.9
dj-database-url>=2.1

# PDF and AI
pdfplumber>=0.10
chromadb>=0.4
openai>=1.0
tiktoken>=0.5

# Testing
pytest>=7.4
pytest-django>=4.5
